Moved test helper(s)
